Implements #710. Adding new unit tests. Conflicts: 	.gitignore 	src/main/java/com/jaeksoft/searchlib/Client.java 	src/main/java/com/jaeksoft/searchlib/index/IndexSingle.java 	src/main/java/com/jaeksoft/searchlib/index/WriterInterface.java 	src/main/java/com/jaeksoft/searchlib/index/WriterLucene.java 	src/main/java/com/jaeksoft/searchlib/result/ResultDocuments.java

+++ src/main/java/com/jaeksoft/searchlib/scheduler/task/TaskDeleteQuery.java
@@ -107,7 +107,7 @@ public class TaskDeleteQuery extends TaskAbstract {
 			request = (AbstractSearchRequest) abstractRequest;
 		}
 		taskLog.setInfo("Deletion request");
-		int i = client.deleteDocuments(request);
+		long i = client.deleteDocuments(request);
 		taskLog.setInfo(i + " document(s) deleted");
 	}
 }
